move the ControlTable to a models package
